+++ src/compiler/csharp_generator.cc
@@ -437,7 +437,7 @@ void GenerateClientStub(Printer* out, const ServiceDescriptor* service,
         "/// <param name=\"channel\">The channel to use to make remote "
         "calls.</param>\n",
         "servicename", GetServiceClassName(service));
-    out->Print("public $name$(grpc::Channel channel) : base(channel)\n", "name",
+    out->Print("public $name$(grpc::ChannelBase channel) : base(channel)\n", "name",
                GetClientClassName(service));
     out->Print("{\n");
     out->Print("}\n");
