+++ src/csharp/Grpc.Reflection/ReflectionGrpc.cs
@@ -69,7 +69,7 @@ namespace Grpc.Reflection.V1Alpha {
     {
       /// <summary>Creates a new client for ServerReflection</summary>
       /// <param name="channel">The channel to use to make remote calls.</param>
-      public ServerReflectionClient(grpc::Channel channel) : base(channel)
+      public ServerReflectionClient(grpc::ChannelBase channel) : base(channel)
       {
       }
       /// <summary>Creates a new client for ServerReflection that uses a custom <c>CallInvoker</c>.</summary>
